adding toasts and stoppers to all records.

+++ src/client/components/Records/ProductsRecords.tsx
@@ -1,5 +1,7 @@
 // Import Dependencies
 import React, { useState, ChangeEvent, useEffect } from 'react';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
@@ -12,6 +14,7 @@ import Input from '@material-ui/core/Input';
 import EditIcon from '@mui/icons-material/Edit';
 import DoneIcon from '@mui/icons-material/Done';
 import DeleteIcon from '@mui/icons-material/Delete';
+import swal from 'sweetalert';
 import axios from 'axios';
 
 interface Column {
@@ -19,7 +22,6 @@ interface Column {
   id: 'id' | 'name' | 'description' | 'harvest_dates';
   label: string;
   minWidth?: number;
-  align?: 'right';
   format?: (value: number) => string;
 }
 
@@ -30,7 +32,6 @@ const columns: readonly Column[] = [
     id: 'description',
     label: 'Description',
     minWidth: 170,
-    align: 'right',
   },
   // {
   //   id: 'subscriptionId',
@@ -48,7 +49,6 @@ const columns: readonly Column[] = [
     id: 'harvest_dates',
     label: 'Harvest Dates',
     minWidth: 170,
-    align: 'right',
   },
 ];
 
@@ -56,6 +56,7 @@ const ProductsRecords = () => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [rows, setRows] = useState([]);
+  const [rowEditing, setRowEditing] = useState(null);
   const [editing, setEditing] = useState(false);
   const [rowColumnId, setRowColumnId] = useState({});
   const [deleteCount, setDeleteCount] = useState(0);
@@ -78,6 +79,15 @@ const ProductsRecords = () => {
         `/api/products/${productId}`,
         updatedProduct,
       );
+      toast.success('Vendor Updated', {
+        position: "top-right",
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        });
       return data;
     } catch (err) {
       console.error(err);
@@ -88,15 +98,31 @@ const ProductsRecords = () => {
   };
 
   const deleteProduct = async (productId: string) => {
-    try {
-      const { data } = await axios.delete(`/api/products/${productId}`);
-      return data;
-    } catch (err) {
-      console.error(err);
-      return {
-        error: err,
-      };
-    }
+    swal({
+      title: 'Are you sure?',
+      text: 'Vendor will be deleted, along with all associated products!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then( async (willDelete) => {
+      if (willDelete) {
+        swal('Event has been deleted', {
+          icon: 'success',
+        });
+        try {
+          const { data } = await axios.delete(`/api/products/${productId}`);
+          setDeleteCount((deleteCount) => deleteCount + 1)
+          return data;
+        } catch (err) {
+          console.error(err);
+          return {
+            error: err,
+          };
+        }
+      } else {
+        swal('That was a close one!');
+      }
+    });
   };
 
   useEffect(() => {
@@ -135,17 +161,29 @@ const ProductsRecords = () => {
     patchProducts(row.id, row);
   };
 
-  const onEdit = () => {
-    setEditing(!editing);
-  };
+  const onEdit = (row: object) => {
+    setEditing(!editing)
+    setRowEditing(row.id)
+  }
 
   const onDelete = (row: object) => {
-    setDeleteCount(deleteCount + 1);
     deleteProduct(row.id);
+    setDeleteCount(deleteCount + 1);
   };
 
   return (
-    <Paper sx={{ width: '100%', overflow: 'hidden' }}>
+    <Paper sx={{ width: '90%', overflow: 'hidden' }}>
+      <ToastContainer
+        position="top-right"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+      />
       <TableContainer sx={{ maxHeight: 440 }}>
         <Table stickyHeader aria-label='sticky table'>
           <TableHead>
@@ -170,7 +208,7 @@ const ProductsRecords = () => {
                     const value = row[column.id];
                     return (
                       <TableCell key={column.id} align={column.align}>
-                        {editing ? (
+                        {(editing && row.id === rowEditing) ? (
                           <Input
                             // type={String}
                             defaultValue={value}
@@ -185,12 +223,12 @@ const ProductsRecords = () => {
                       </TableCell>
                     );
                   })}
-                  {editing ? (
+                  {(editing && row.id === rowEditing) ? (
                     <DoneIcon onClick={() => onDone(row)} />
                   ) : (
                     <>
                       <TableCell>
-                        <EditIcon onClick={() => onEdit()} />
+                        <EditIcon onClick={() => onEdit(row)} />
                       </TableCell>
                       <TableCell>
                         <DeleteIcon onClick={() => onDelete(row)} />
