+++ src/client/components/Records/SubscriptionEntriesRecords.tsx
@@ -1,5 +1,7 @@
 // Import Dependencies
 import React, { useState, ChangeEvent, useEffect } from 'react';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
@@ -8,17 +10,17 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import Input from "@material-ui/core/Input";
+import Input from '@material-ui/core/Input';
 import EditIcon from '@mui/icons-material/Edit';
 import DoneIcon from '@mui/icons-material/Done';
 import DeleteIcon from '@mui/icons-material/Delete';
+import swal from 'sweetalert';
 import axios from 'axios';
 
 interface Column {
   id: 'id' | 'userId' | 'subscriptionId';
   label: string;
   minWidth?: number;
-  align?: 'right';
   format?: (value: number) => string;
 }
 
@@ -28,7 +30,6 @@ const columns: readonly Column[] = [
     id: 'subscriptionId',
     label: 'Subscription ID',
     minWidth: 170,
-    align: 'right',
   },
 ];
 
@@ -36,6 +37,7 @@ const SubscriptionEntriesRecords = () => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [rows, setRows] = useState([]);
+  const [rowEditing, setRowEditing] = useState(null);
   const [editing, setEditing] = useState(false);
   const [rowColumnId, setRowColumnId] = useState({});
   const [deleteCount, setDeleteCount] = useState(0);
@@ -55,6 +57,15 @@ const SubscriptionEntriesRecords = () => {
   const patchSubscriptionEntries = async (subscriptionEntrieId: string, updatedSubscriptionEntrie: any) => {
     try {
       const { data } = await axios.patch(`/api/subscription-entries/${subscriptionEntrieId}`, updatedSubscriptionEntrie);
+      toast.success('Vendor Updated', {
+        position: "top-right",
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        });
       return data
     } catch (err) {
       console.error(err)
@@ -65,15 +76,31 @@ const SubscriptionEntriesRecords = () => {
   }
 
   const deleteSubscriptionEntries = async (subscriptionEntrieId: string) => {
-    try {
-      const { data } = await axios.delete(`/api/subscription-entries/${subscriptionEntrieId}`);
-      return data
-    } catch (err) {
-      console.error(err)
-      return {
-        error: err
+    swal({
+      title: 'Are you sure?',
+      text: 'Vendor will be deleted, along with all associated products!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then( async (willDelete) => {
+      if (willDelete) {
+        swal('Event has been deleted', {
+          icon: 'success',
+        });
+        try {
+          const { data } = await axios.delete(`/api/subscription-entries/${subscriptionEntrieId}`);
+          setDeleteCount((deleteCount) => deleteCount + 1)
+          return data;
+        } catch (err) {
+          console.error(err);
+          return {
+            error: err,
+          };
+        }
+      } else {
+        swal('That was a close one!');
       }
-    }
+    });
   }
 
   useEffect(() => {
@@ -112,17 +139,29 @@ const SubscriptionEntriesRecords = () => {
     patchSubscriptionEntries(row.id, row);
   }
 
-  const onEdit = () => {
+  const onEdit = (row: object) => {
     setEditing(!editing)
+    setRowEditing(row.id)
   }
 
   const onDelete = (row: object) => {
-    setDeleteCount(deleteCount + 1)
     deleteSubscriptionEntries(row.id)
+    setDeleteCount(deleteCount + 1)
   }
 
   return (
-    <Paper sx={{ width: '100%', overflow: 'hidden' }}>
+    <Paper sx={{ width: '90%', overflow: 'hidden' }}>
+      <ToastContainer
+        position="top-right"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+      />
       <TableContainer sx={{ maxHeight: 440 }}>
         <Table stickyHeader aria-label='sticky table'>
           <TableHead>
@@ -147,7 +186,7 @@ const SubscriptionEntriesRecords = () => {
                     const value = row[column.id];
                     return (
                       <TableCell key={column.id} align={column.align}>
-                        {editing ? (
+                        {(editing && row.id === rowEditing) ? (
                           <Input
                             // type={String}
                             defaultValue={value}
@@ -161,19 +200,18 @@ const SubscriptionEntriesRecords = () => {
                       </TableCell>
                     );
                   })}
-                  {editing ? (
+                  {(editing && row.id === rowEditing) ? (
                     <DoneIcon onClick={() => onDone(row)} />
-                  ) :
+                  ) : (
                     <>
                       <TableCell>
-                        <EditIcon onClick={() => onEdit()} />
+                        <EditIcon onClick={() => onEdit(row)} />
                       </TableCell>
                       <TableCell>
                         <DeleteIcon onClick={() => onDelete(row)} />
                       </TableCell>
                     </>
-                  }
-
+                  )}
                 </TableRow>
               ))}
           </TableBody>
